+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 - feat(utils): add `"soft"` option to `Powerset.to_multilabel`
 - improve(pipeline): compute `fbank` on GPU when requested
 - fix(pipeline): fix `AgglomerativeClustering` to honor `num_clusters` when provided
+- BREAKING(pipeline): rename `WeSpeakerPretrainedSpeakerEmbedding` to `ONNXWeSpeakerPretrainedSpeakerEmbedding`
+- BREAKING(setup): remove `onnxruntime` dependency.
+  You can still use ONNX `hbredin/wespeaker-voxceleb-resnet34-LM` but you will have to install `onnxruntime` yourself.
 
 ## Version 3.0.1 (2023-09-28)